+++ docs/articles/release_notes.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 * Added [RelaxModifier](modifiers/relaxmodifier.md)
-* Added [PlanarLazyGrid](grids/planarlazygrid.md)
+* Added [PlanarLazyMeshGrid](grids/planarlazymeshgrid.md)
 * Added [Townscaper grid](xref:Sylves.TownscaperGrid)
 * Added [Townscaper tutorial](tutorials/townscaper.md)
 * Added [Concat](xref:Sylves.MeshDataOperations.Concat(System.Collections.Generic.IEnumerable{Sylves.MeshData},System.Collections.Generic.List{System.Int32[]}@)), [Relax](xref:Sylves.MeshDataOperations.Relax(Sylves.MeshData,System.Int32)), [RandomPairing](xref:Sylves.MeshDataOperations.RandomPairing(Sylves.MeshData,System.Func{System.Double})), [MaxRandomPairing](xref:Sylves.MeshDataOperations.MaxRandomPairing(Sylves.MeshData,System.Func{System.Double})) and [Weld](xref:Sylves.MeshDataOperations.Weld(Sylves.MeshData,System.Single)) mesh operations.